Revert "Attempt to add builtin_defs to CLI"
This reverts commit 2679077de91e32842c04967bbad4ae101688e2e3.
